✨  - Moved profile picture; fixed up headers
This was a combination of a branch @SaraSweetie was working on earlier; and some other bugs... primary purpose was to move the user's profile picture into the top right corner. The hamburger button is currently not working but all the buttons have been place in the header for testing purposes #40
